+++ golang/testsuite/testsuite_impl/example_testsuite.go
@@ -10,6 +10,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl/advanced_network_test"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl/basic_datastore_and_api_test"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl/basic_datastore_test"
+	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl/bulk_command_execution_test"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl/exec_command_test"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl/files_artifact_mounting_test"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl/network_partition_test"
@@ -49,6 +50,7 @@ func (suite ExampleTestsuite) GetTests() map[string]testsuite.Test {
 		)
 		tests["filesArtifactMountingTest"] = files_artifact_mounting_test.FilesArtifactMountingTest{}
 		tests["execCommandTest"] = exec_command_test.ExecCommandTest{}
+		tests["bulkCommandExecutionTest"] = bulk_command_execution_test.NewBulkCommandExecutionTest(suite.datastoreServiceImage)
 	}
 
 	return tests